Translate unicode to ASCII, decrease JSON error wait time
The unicode used in HQ's questions and answers were messing up the search process. An easy solution was to translate all questions, answers, and search results into ASCII so answers would be easily searchable.

The 10 second JSON error sleep time in hq_main made the program fail to recognize a question, so a 1 second wait should be sufficient.